+++ test/Libplanet.Net.Tests/SwarmTest.Preload.cs
@@ -136,9 +136,9 @@ namespace Libplanet.Net.Tests
                 Block block = minerChain.EvaluateAndSign(
                     ProposeNext(
                         previousBlock: i == 0 ? minerChain.Genesis : blocks[i - 1],
-                        miner: ChainPrivateKey.PublicKey,
+                        miner: GenesisProposer.PublicKey,
                         lastCommit: CreateBlockCommit(minerChain.Tip)),
-                    ChainPrivateKey);
+                    GenesisProposer);
                 blocks.Add(block);
                 if (i != 11)
                 {
@@ -481,7 +481,7 @@ namespace Libplanet.Net.Tests
                     DateTimeOffset.UtcNow);
 
                 Block block = minerChain.ProposeBlock(
-                    ChainPrivateKey,
+                    GenesisProposer,
                     new[] { tx }.ToImmutableList(),
                     CreateBlockCommit(minerChain.Tip),
                     ImmutableArray<EvidenceBase>.Empty);
@@ -995,7 +995,7 @@ namespace Libplanet.Net.Tests
             minerChain1.Append(block2, CreateBlockCommit(block2));
 
             Block block = minerChain2.ProposeBlock(
-                ChainPrivateKey, CreateBlockCommit(minerChain2.Tip));
+                GenesisProposer, CreateBlockCommit(minerChain2.Tip));
             minerChain2.Append(block, CreateBlockCommit(block));
 
             Assert.True(minerChain1.Count > minerChain2.Count);