+++ test/Libplanet.Tests/Blocks/BlockFixture.cs
@@ -1,6 +1,5 @@
 using System.Collections.Generic;
 using System.Collections.Immutable;
-using System.Numerics;
 using System.Security.Cryptography;
 using Libplanet.Common;
 using Libplanet.Crypto;
@@ -57,7 +56,7 @@ namespace Libplanet.Tests.Blocks
                             Next.Hash,
                             Next.Timestamp,
                             Miner.PublicKey,
-                            BigInteger.One,
+                            TestUtils.ValidatorSet.GetValidator(Miner.PublicKey).Power,
                             VoteFlag.PreCommit).Sign(Miner),
                     }.ToImmutableArray()),
                 evidence: ImmutableArray<EvidenceBase>.Empty