+++ CHANGES.md
@@ -10,11 +10,26 @@ To be released.
 
 ### Backward-incompatible API changes
 
+ -  (Libplanet.Action) Added `MaxGasPrice` property to `IActionContext`
+    interface and its implementations.  [[#3762]]
+ -  (Libplanet.Action) Removed `IFeeCollector` interface
+    and its implementations.  [[#3867]]
+ -  (Libplanet.Action) Removed following methods from the
+    `IActionContext` interface.  [[#3868]]
+     -  Removed `IActionContext.UseGas(long)`.
+     -  Removed `IActionContext.GasUsed()`.
+     -  Removed `IActionContext.GasLimit()`.
+ -  (Libplanet.Action) Added `LastCommit` property to `IActionContext`
+    interface and its implementations.  [[#3912]]
+ -  (Libplanet.Explorer) Added `self` field to `NoteStateType`.  [[#3841]]
+
 ### Backward-incompatible network protocol changes
 
 ### Backward-incompatible storage format changes
 
 ### Added APIs
+
+ -  (Libplanet.Action) Added `GasTracer` static class.  [[#3868]]
 
 ### Behavioral changes
 
@@ -23,6 +38,12 @@ To be released.
 ### Dependencies
 
 ### CLI tools
+
+[#3762]: https://github.com/planetarium/libplanet/pull/3762
+[#3841]: https://github.com/planetarium/libplanet/pull/3841
+[#3867]: https://github.com/planetarium/libplanet/pull/3867
+[#3868]: https://github.com/planetarium/libplanet/pull/3868
+[#3912]: https://github.com/planetarium/libplanet/pull/3912
 
 
 Version 5.2.0